### 2024-10-31 #7418 @editorialbot
[PRE REVIEW]: HOHQMesh: An All Quadrilateral/Hexahedral Unstructured Mesh Generator for High Order Elements
**Submitting author:** @andrewwinters5000 (<a href="http://orcid.org/0000-0002-5902-1522">Andrew Winters</a>)
**Repository:** https://github.com/trixi-framework/HOHQMesh
**Branch with paper.md** (empty if default branch): paper-2024-joss
**Version:** v1.5.2
