+++ paper.bib
@@ -1,10 +1,10 @@
 %% This BibTeX bibliography file was created using BibDesk.
 %% https://bibdesk.sourceforge.io/
 
-%% Created for David Kopriva at 2024-10-29 13:34:44 -0700 
+%% Created for David Kopriva at 2024-10-29 13:34:44 -0700
 
 
-%% Saved with string encoding Unicode (UTF-8) 
+%% Saved with string encoding Unicode (UTF-8)
 
 
 
@@ -72,8 +72,7 @@
 	author = {Sherwin, S. J. and Peir{\'o}, J.},
 	date-added = {2024-10-14 17:04:29 -0700},
 	date-modified = {2024-10-14 17:04:39 -0700},
-	doi = {https://doi.org/10.1002/nme.397},
-	eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1002/nme.397},
+	doi = {10.1002/nme.397},
 	journal = {International Journal for Numerical Methods in Engineering},
 	keywords = {high-order surface representation, hybrid meshing, curvature based adaption, high-order elements},
 	number = {1},
@@ -81,9 +80,7 @@
 	title = {Mesh generation in curvilinear domains using high-order elements},
 	url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/nme.397},
 	volume = {53},
-	year = {2002},
-	bdsk-url-1 = {https://onlinelibrary.wiley.com/doi/abs/10.1002/nme.397},
-	bdsk-url-2 = {https://doi.org/10.1002/nme.397}}
+	year = {2002}}
 
 @incollection{HOPR,
 	author = {F. Hindenlang and T. Bolemann and C-D. Munz},
@@ -111,7 +108,7 @@
 	author = {Hervouet, Jean-Michel and Petitjean, Alain},
 	date-added = {2024-10-14 13:59:15 -0700},
 	date-modified = {2024-10-14 13:59:15 -0700},
-	doi = {0.1080/00221689909498511},
+	doi = {10.1080/00221689909498511},
 	journal = {Journal of hydraulic research},
 	number = {6},
 	pages = {777--788},
@@ -119,7 +116,7 @@
 	title = {Malpasset dam-break revisited with two-dimensional computations},
 	volume = {37},
 	year = {1999},
-	bdsk-url-1 = {https://doi.org/0.1080/00221689909498511}}
+	bdsk-url-1 = {https://doi.org/10.1080/00221689909498511}}
 
 @article{Bommes:2013lc,
 	abstract = {Triangle meshes have been nearly ubiquitous in computer graphics, and a large body of data structures and geometry processing algorithms based on them has been developed in the literature. At the same time, quadrilateral meshes, especially semi-regular ones, have advantages for many applications, and significant progress was made in quadrilateral mesh generation and processing during the last several years. In this survey we discuss the advantages and problems of techniques operating on quadrilateral meshes, including surface analysis and mesh quality, simplification, adaptive refinement, alignment with features, parametrisation and remeshing.},
@@ -172,7 +169,7 @@ Program title: Nektar++ Catalogue identifier: AEVV_v1_0 Program summary URL:http
 	author = {C.D. Cantwell and D. Moxey and A. Comerford and A. Bolis and G. Rocco and G. Mengaldo and D. {De Grazia} and S. Yakovlev and J.-E. Lombard and D. Ekelschot and B. Jordi and H. Xu and Y. Mohamied and C. Eskilsson and B. Nelson and P. Vos and C. Biotto and R.M. Kirby and S.J. Sherwin},
[…]
