### 2024-10-31 #7418 @andrewwinters5000
> 👋 @andrewwinters5000 - Thanks for your submission. I'll try to find an editor next, though I may be delayed a bit over next week.
>
> While I do so, you could work on the invalid DOIs that editorialbot found. One seems to be missing a "1" at the start, and the others have prefixes that should be deleted. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` when the references are right to make a new PDF. editorialbot commands need to be the first entry in a new comment.

Thanks! I have fixed the broken DOIs and will suggest reviewers as instructed above.
