### 2024-11-04 de17f82
Removed initDefaultEllipse subroutine
