### 2024-11-04 205f4e7
Added printDescription tests for line and ellipse
