+++ Source/Curves/ContinuousCurves/SMEllipticArc.f90
@@ -197,7 +197,8 @@
             IMPLICIT NONE
             CLASS(SMEllipticArc) :: self
             INTEGER              :: iUnit
-            WRITE(iUnit,*) "SMEllipticArc Object"
+            OPEN(iUnit, STATUS='SCRATCH')
+            WRITE(iUnit,'(A)') "SMEllipticArc Object"
             IF(self % refCount() >= 0) CONTINUE
         END SUBROUTINE printEllipticArcDescription
 !
