+++ Source/Testing/CurveTests.f90
@@ -52,6 +52,7 @@
          REAL(KIND=RP)                           :: xExact(3)
          INTEGER                                 :: j
          REAL(KIND=RP)                           :: dt = 0.2_RP, t, e, tol = 1.0d-5
+         INTEGER                                 :: iUnit = 200
 !
 !        ---------------
 !        Line definition
@@ -84,7 +85,6 @@
 !
          REAL(KIND=RP)       :: r(7), xA(7), yA(7), zA(7)
          TYPE(SMSplineCurve) :: spline
-
 !
 !        ---------
 !        Test Line
@@ -109,6 +109,8 @@
          
          CALL FTAssert(test = line % curveIsStraight(), msg = "Line straightness")
          CALL destructLine(self = line)
+         CALL printLineDescription(self = line, iUnit = iUnit)
+         CALL TestPrintDescription("SMLine Object", iUnit)
 !
 !        -----------
 !        Test circle
@@ -166,6 +168,8 @@
          END DO
          CALL FTAssert(test = .NOT. ellipse % curveIsStraight(), msg = "Ellipse shouldn't be straight")
          CALL destructEllipticArc(self = ellipse)
+         CALL printEllipticArcDescription(self = ellipse, iUnit = iUnit)
+         CALL TestPrintDescription("SMEllipticArc Object", iUnit)
 !
 !        ------------------------------
 !        Parametric equation curve test
@@ -295,8 +299,40 @@
         CALL ExactAndComputedQuadraticHessian(x = 0.5_RP,y = 0.5_RP,exact = exactHess,computed = hessF)
         CALL FTAssertEqual(expectedValue = 0.0_RP,                       &
                            actualValue   = MAXVAL(ABS(exactHess-hessF)), &
-                           relTol         = 2.0d-8,                      &
-                           absTol         = 1.0d-9,                      &
+                           relTol        = 2.0d-8,                       &
+                           absTol        = 1.0d-9,                       &
                            msg           = "Quadratic Hessian error")
 
       END SUBROUTINE TestGaussianCurvature
+!
+!     ---------------------
+!     Test printDescription
+!     ---------------------
+!
+      SUBROUTINE TestPrintDescription(description, iUnit)
+         USE FTAssertions
+         IMPLICIT NONE
+         INTEGER            :: diffCheck = 0
+         INTEGER            :: iUnit, k
+         CHARACTER(len=*)   :: description
+         CHARACTER(len=100) :: fileLine
+
+         REWIND(iUnit)
+         READ(iUnit, '(A)') fileLine
+         IF (LEN_TRIM(fileLine) /= LEN_TRIM(description)) THEN
+            diffCheck = diffCheck + 1
+         ELSE
+            DO k = 1, MIN(LEN_TRIM(fileLine), LEN_TRIM(description))
+               IF (fileLine(k:k) /= description(k:k)) THEN
+                  diffCheck = diffCheck + 1
+               END IF
+            END DO
+         END IF
+         CALL FTAssertEqual(expectedValue = 0,                                                 &
+                            actualValue   = diffCheck,                                         &
+                            msg           = "Description not writing correctly; should be:" // &
+                            ACHAR(10) // "       " // description)
+         diffCheck = 0
+         iUnit = iUnit + 1
+
+      END SUBROUTINE TestPrintDescription
