### 2024-11-06 a285dd8
Modify Curve tests
Move the scratch iUnit open to test routine out of the printDescription procedure.
