### 2024-11-06 9d7b811
Update CurveTests.f90
Use newunit to generate an unused file unit. Remove extraneous lines in TestPrintDescription
