### 2024-11-06 f498508
Update CurveTests.f90
Fix SAVEd bug introduced in last commit.
