### 2024-11-07 2effb0b
AddGallery
Add a Gallery page with plots.
