+++ Documentation/docs/Gallery.md
@@ -1,0 +1,99 @@
+<!---
+Interesting meshes can be added below, with new categories added as needed. A two-column
+format is used, so to add two plots, use
+
+<div align="center">
+  <img src="hohqmesh-gallery/meshPlotLeft.png" width="48%" />
+   [Contributor. Description]: #
+ <img src="hohqmesh-gallery/meshPlotright.png" width="48%" /> 
+   [Contributor. Description]: #
+</div>
+
+and for a single,
+<div align="left">
+  <img src="hohqmesh-gallery/meshPlot.png" width="48%" />
+   [Contributor. Description]: #
+</div>
+
+The Contributor is who submitted the mesh. For now Contributor. Description are in comments. If someone knows how to add captions to the plots, and acknowlegements for
+who contributed them, that would be nice to have those added and use these for the caption.
+
+At some point these figures should be moved out of the repo to where the other figures are located.
+
+-->
+
+# A HOHQMesh Mesh Gallery
+
+Please add interesting contributions!
+
+## Exterior Domains
+
+<div align="center">
+  <img src="hohqmesh-gallery/cylinder-flow.png" width="48%" />
+  [Contributor?. Cylinder with a wake region]: #
+  <img src="hohqmesh-gallery/naca0012.png" width="48%" /> 
+  [David A. Kopriva. NACA0012 Airfoil]:
+</div>
+
+<div align="center">
+  <img src="hohqmesh-gallery/three-element-30p30n.png" width="48%" />
+   [Andrew R. Winters. Three element Airfoil from http://www.ae.metu.edu.tr/tuncer/ae546/multi/]: #
+<img src="hohqmesh-gallery/naca6412.png" width="48%" /> 
+  [Andrew R. Winters. NACA6412 Airfoil from https://github.com/cfsengineering/GMSH-Airfoil-2D]: #
+</div>
+
+## Coastlines
+
+<div align="center">
+  <img src="hohqmesh-gallery/indian-ocean.png" width="48%" />
+  [David A. Kopriva. Indian Ocean]: #  
+  <img src="hohqmesh-gallery/SuperiorFull.png" width="48%" /> 
+  [David A. Kopriva. Lake Superior]: #  
+</div>
+<div align="center">
+  <img src="hohqmesh-gallery/east-coast-us-2d.png" width="48%" />
+  [Joe Schoonover. East coast of the US]: #  
+  <img src="hohqmesh-gallery/malpasset.png" width="48%" /> 
+  [Andrew R. Winters.  Reyran river valley down to the Mediterranean sea]: #  
+</div>
+
+## Embedded
+
+<div align="left">
+  <img src="hohqmesh-gallery/embedded-boundaries.png" width="48%" />
+  [David A. Kopriva. Three embedded circles]: #
+</div>
+
+## Topography
+
+<div align="center">
+  <img src="hohqmesh-gallery/mt-st-helens-above.png" width="48%" />
+  [Andrew R. Winters.  Mount St. Helens from above]: #  
+  <img src="hohqmesh-gallery/mt-st-helens-perspective.png" width="48%" /> 
+  [Andrew R. Winters.  Mount St. Helens, perspective view]: #  
+</div>
+<div align="left">
+  <img src="hohqmesh-gallery/east-coast-us-3d.png" width="48%" />
+  [Joe Schoonover. East coast of the US with topography]: #  
+</div>
+
+## Just for Fun
+
+<div align="center">
+  <img src="hohqmesh-gallery/trixi-hexe.png" width="48%" />
+  [Andrew R. Winters. Trixi logo]: #  
[…]
