### 2024-11-07 94a985f
Update Gallery.md
Fix misspelling
