### 2024-11-08 85e586d
Add EllipticArc Module (#112)
* Added SMEllipticArcClass module

* Added documentation for elliptic curve

* Fixed minor syntax errors

* Shortened warning message to suppress "Error: Line truncated"

* Fixed typos

* Added RotatedEllipseAnd4Circle example

* Fixed more typos

* Added initWithParametersNameAndID_SMEllipticArc override

* Fixed typos

* Corrected documentation

* Changed RotatedEllipse example for coverage

* Added elliptic arc to AllFeatures.control

* Added RotatedEllipse example to Benchmarks

* Updated documentation

* Update Benchmarks/TestGeneratorControlFiles/AllFeatures
[…]
