+++ Source/Curves/ContinuousCurves/SMCircularArc.f90
@@ -1,198 +1,0 @@
-! MIT License
-!
-! Copyright (c) 2010-present David A. Kopriva and other contributors: AUTHORS.md
-!
-! Permission is hereby granted, free of charge, to any person obtaining a copy  
-! of this software and associated documentation files (the "Software"), to deal  
-! in the Software without restriction, including without limitation the rights  
-! to use, copy, modify, merge, publish, distribute, sublicense, and/or sell  
-! copies of the Software, and to permit persons to whom the Software is  
-! furnished to do so, subject to the following conditions:
-!
-! The above copyright notice and this permission notice shall be included in all  
-! copies or substantial portions of the Software.
-!
-! THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR  
-! IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,  
-! FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE  
-! AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER  
-! LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,  
-! OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE  
-! SOFTWARE.
-!
-! --- End License
-!
-!////////////////////////////////////////////////////////////////////////
-!
-!      SMCircularArc.f90
-!      Created: July 30, 2013 4:04 PM 
-!      By: David Kopriva  
-!
-!////////////////////////////////////////////////////////////////////////
-!
-      Module SMCircularArcClass
-      USE SMCurveClass
-      USE SMConstants
-      USE ProgramGlobals
-      IMPLICIT NONE
-!
-!     ---------------------
-!     Class type definition
-!     ---------------------
-!
-      TYPE, EXTENDS(SMCurve) :: SMCircularArc
-         REAL(KIND=RP) :: center(3) 
-         REAL(KIND=RP) :: radius
-         REAL(KIND=RP) :: startAngle
-         REAL(KIND=RP) :: endAngle
-!
-!        ========
-         CONTAINS
-!        ========
-!
-         PROCEDURE :: initWithParametersNameAndID
-         FINAL     :: destructCircularArc
-         PROCEDURE :: positionAt       => positionOnCircularArcAt
-         PROCEDURE :: tangentAt        => tangentOnCircularArcAt
-         PROCEDURE :: printDescription => printCircularArcDescription
-         PROCEDURE :: className        => ArcClassName
-      END TYPE SMCircularArc
-      
-      CHARACTER(LEN=STRING_CONSTANT_LENGTH), PARAMETER :: CIRCULAR_ARC_CONTROL_KEY      = "CIRCULAR_ARC"
-      CHARACTER(LEN=STRING_CONSTANT_LENGTH), PARAMETER :: CIRCULAR_ARC_RADIUS_KEY       = "radius"
-      CHARACTER(LEN=STRING_CONSTANT_LENGTH), PARAMETER :: CIRCULAR_ARC_CENTER_KEY       = "center"
-      CHARACTER(LEN=STRING_CONSTANT_LENGTH), PARAMETER :: CIRCULAR_ARC_UNITS_KEY        = "units"
[…]
