+++ Source/Curves/ContinuousCurves/SMEllipticArc.f90
@@ -1,0 +1,223 @@
+! MIT License
+!
+! Copyright (c) 2010-present David A. Kopriva and other contributors: AUTHORS.md
+!
+! Permission is hereby granted, free of charge, to any person obtaining a copy  
+! of this software and associated documentation files (the "Software"), to deal  
+! in the Software without restriction, including without limitation the rights  
+! to use, copy, modify, merge, publish, distribute, sublicense, and/or sell  
+! copies of the Software, and to permit persons to whom the Software is  
+! furnished to do so, subject to the following conditions:
+!
+! The above copyright notice and this permission notice shall be included in all  
+! copies or substantial portions of the Software.
+!
+! THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR  
+! IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,  
+! FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE  
+! AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER  
+! LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,  
+! OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE  
+! SOFTWARE.
+!
+! --- End License
+!
+!////////////////////////////////////////////////////////////////////////
+!
+!      SMEllipticArc.f90
+!      Created: October 22, 2024 9:42 PM 
+!      By: Garrett Byrd
+!
+!////////////////////////////////////////////////////////////////////////
+!
+        Module SMEllipticArcClass
+        USE SMCurveClass
+        USE SMConstants
+        USE ProgramGlobals
+        IMPLICIT NONE
+!
+!     ---------------------
+!     Class type definition
+!     ---------------------
+!
+        TYPE, EXTENDS(SMCurve) :: SMEllipticArc
+            REAL(KIND=RP) :: center(3)
+            REAL(KIND=RP) :: xRadius
+            REAL(KIND=RP) :: yRadius
+            REAL(KIND=RP) :: startAngle
+            REAL(KIND=RP) :: endAngle
+            REAL(KIND=RP) :: rotation
+!
+!           ========
+            CONTAINS
+!           ========
+!
+            GENERIC :: initWithParametersNameAndID  => initWithParametersNameAndID_SMEllipticArc, &
+                                                       initWithParametersNameAndID_SMCircularArc
+            PROCEDURE, PRIVATE :: initWithParametersNameAndID_SMEllipticArc
+            PROCEDURE, PRIVATE :: initWithParametersNameAndID_SMCircularArc
+            FINAL     :: destructEllipticArc
+            PROCEDURE :: positionAt       => positionOnEllipticArcAt
+            PROCEDURE :: tangentAt        => tangentOnEllipticArcAt
+            PROCEDURE :: printDescription => printEllipticArcDescription
+            PROCEDURE :: className        => EllipticArcClassName
+        END TYPE SMEllipticArc
+
+        CHARACTER(LEN=STRING_CONSTANT_LENGTH), PARAMETER :: ELLIPTIC_ARC_CONTROL_KEY      = "ELLIPTIC_ARC"
+        CHARACTER(LEN=STRING_CONSTANT_LENGTH), PARAMETER :: ELLIPTIC_ARC_X_RADIUS_KEY     = "xRadius"
[…]
