+++ Benchmarks/BenchmarkData/RotatedEllipseWithSubregions.txt
@@ -1,0 +1,9 @@
+           1
+        2436        2674        5116           7
+ 8.99660E-03
+ 1.09943E+00
+ 1.07198E+00
+ 1.19205E+00
+ 8.43429E-03
+ 8.30762E+01
+ 9.80230E+01

[…]
