### 2024-11-08 8534537
GalleryReferences
Add references to the gallery in the README and to the menu that shows up on the left. I also rearranged the menu to have the authors and license at the end.
