### 2024-11-09 cfc6e20
Update README.md
Update documentation to include new elliptic arc and Malpasset examples.
