### 2024-11-11 7900c4a
use gallery images from examples repo
