+++ Documentation/docs/Gallery.md
@@ -3,15 +3,15 @@ Interesting meshes can be added below, with new categories added as needed. A tw
 format is used, so to add two plots, use
 
 <div align="center">
-  <img src="hohqmesh-gallery/meshPlotLeft.png" width="48%" />
+  <img src="https://github.com/trixi-framework/HOHQMesh_examples/blob/9f4996ce1f3cc94f2dc0a537135c8d696e61de20/gallery/meshPlotLeft.png" width="48%" />
    [Contributor. Description]: #
- <img src="hohqmesh-gallery/meshPlotright.png" width="48%" /> 
+ <img src="https://github.com/trixi-framework/HOHQMesh_examples/blob/9f4996ce1f3cc94f2dc0a537135c8d696e61de20/gallery/meshPlotright.png" width="48%" />
    [Contributor. Description]: #
 </div>
 
 and for a single,
 <div align="left">
-  <img src="hohqmesh-gallery/meshPlot.png" width="48%" />
+  <img src="https://github.com/trixi-framework/HOHQMesh_examples/blob/9f4996ce1f3cc94f2dc0a537135c8d696e61de20/gallery/meshPlot.png" width="48%" />
    [Contributor. Description]: #
 </div>
 
@@ -29,71 +29,71 @@ Please submit interesting contributions!
 ## Exterior Domains
 
 <div align="center">
-  <img src="hohqmesh-gallery/cylinder-flow.png" width="48%" />
+  <img src="https://github.com/trixi-framework/HOHQMesh_examples/blob/9f4996ce1f3cc94f2dc0a537135c8d696e61de20/gallery/cylinder-flow.png?raw=true" width="48%" />
   [Contributor?. Cylinder with a wake region]: #
-  <img src="hohqmesh-gallery/naca0012.png" width="48%" /> 
+  <img src="https://github.com/trixi-framework/HOHQMesh_examples/blob/9f4996ce1f3cc94f2dc0a537135c8d696e61de20/gallery/naca0012.png?raw=true" width="48%" />
   [David A. Kopriva. NACA0012 Airfoil]:
 </div>
 
 <div align="center">
-  <img src="hohqmesh-gallery/three-element-30p30n.png" width="48%" />
+  <img src="https://github.com/trixi-framework/HOHQMesh_examples/blob/9f4996ce1f3cc94f2dc0a537135c8d696e61de20/gallery/three-element-30p30n.png?raw=true" width="48%" />
    [Andrew R. Winters. Three element Airfoil from http://www.ae.metu.edu.tr/tuncer/ae546/multi/]: #
-<img src="hohqmesh-gallery/naca6412.png" width="48%" /> 
+<img src="https://github.com/trixi-framework/HOHQMesh_examples/blob/9f4996ce1f3cc94f2dc0a537135c8d696e61de20/gallery/naca6412.png?raw=true" width="48%" />
   [Andrew R. Winters. NACA6412 Airfoil from https://github.com/cfsengineering/GMSH-Airfoil-2D]: #
 </div>
 
 ## Coastlines
 
 <div align="center">
-  <img src="hohqmesh-gallery/indian-ocean.png" width="48%" />
-  [David A. Kopriva. Indian Ocean]: #  
-  <img src="hohqmesh-gallery/SuperiorFull.png" width="48%" /> 
-  [David A. Kopriva. Lake Superior]: #  
+  <img src="https://github.com/trixi-framework/HOHQMesh_examples/blob/9f4996ce1f3cc94f2dc0a537135c8d696e61de20/gallery/indian-ocean.png?raw=true" width="48%" />
+  [David A. Kopriva. Indian Ocean]: #
[…]
