### 2024-11-11 54ab647
use smaller versions of the images to decrease latency
