+++ Documentation/docs/Gallery.md
@@ -3,15 +3,15 @@ Interesting meshes can be added below, with new categories added as needed. A tw
 format is used, so to add two plots, use
 
 <div align="center">
-  <img src="https://github.com/trixi-framework/HOHQMesh_examples/blob/9f4996ce1f3cc94f2dc0a537135c8d696e61de20/gallery/meshPlotLeft.png" width="48%" />
+  <img src="https://github.com/trixi-framework/HOHQMesh_examples/blob/a8610caf1c5047741ee341462cb3e2f9d34471c6/gallery/meshPlotLeft.png" width="48%" />
    [Contributor. Description]: #
- <img src="https://github.com/trixi-framework/HOHQMesh_examples/blob/9f4996ce1f3cc94f2dc0a537135c8d696e61de20/gallery/meshPlotright.png" width="48%" />
+ <img src="https://github.com/trixi-framework/HOHQMesh_examples/blob/a8610caf1c5047741ee341462cb3e2f9d34471c6/gallery/meshPlotright.png" width="48%" />
    [Contributor. Description]: #
 </div>
 
 and for a single,
 <div align="left">
-  <img src="https://github.com/trixi-framework/HOHQMesh_examples/blob/9f4996ce1f3cc94f2dc0a537135c8d696e61de20/gallery/meshPlot.png" width="48%" />
+  <img src="https://github.com/trixi-framework/HOHQMesh_examples/blob/a8610caf1c5047741ee341462cb3e2f9d34471c6/gallery/meshPlot.png" width="48%" />
    [Contributor. Description]: #
 </div>
 
@@ -29,78 +29,78 @@ Please submit interesting contributions!
 ## Exterior Domains
 
 <div align="center">
-  <img src="https://github.com/trixi-framework/HOHQMesh_examples/blob/9f4996ce1f3cc94f2dc0a537135c8d696e61de20/gallery/cylinder-flow.png?raw=true" width="48%" />
+  <img src="https://github.com/trixi-framework/HOHQMesh_examples/blob/9df4021e865fcdb600a8a25ce2a37938a1e99523/gallery/cylinder-flow-small.png?raw=true" width="48%" />
   [Contributor?. Cylinder with a wake region]: #
-  <img src="https://github.com/trixi-framework/HOHQMesh_examples/blob/9f4996ce1f3cc94f2dc0a537135c8d696e61de20/gallery/naca0012.png?raw=true" width="48%" />
+  <img src="https://github.com/trixi-framework/HOHQMesh_examples/blob/a8610caf1c5047741ee341462cb3e2f9d34471c6/gallery/naca0012-small.png?raw=true" width="48%" />
   [David A. Kopriva. NACA0012 Airfoil]:
 </div>
 
 <div align="center">
-  <img src="https://github.com/trixi-framework/HOHQMesh_examples/blob/9f4996ce1f3cc94f2dc0a537135c8d696e61de20/gallery/three-element-30p30n.png?raw=true" width="48%" />
+  <img src="https://github.com/trixi-framework/HOHQMesh_examples/blob/a8610caf1c5047741ee341462cb3e2f9d34471c6/gallery/three-element-30p30n-small.png?raw=true" width="48%" />
    [Andrew R. Winters. Three element Airfoil from http://www.ae.metu.edu.tr/tuncer/ae546/multi/]: #
-<img src="https://github.com/trixi-framework/HOHQMesh_examples/blob/9f4996ce1f3cc94f2dc0a537135c8d696e61de20/gallery/naca6412.png?raw=true" width="48%" />
[…]
