### 2024-11-11 #7418 @danielskatz
👋 @HaoZeke - As mentioned in https://github.com/openjournals/joss-reviews/issues/7418#issuecomment-2450619565, would you be able to edit this submission?
