### 2024-11-12 49854b9
Update Gallery.md
Fix multiple the column issue by including html comment markers in for the comments in the div's.
