+++ Documentation/docs/Gallery.md
@@ -16,7 +16,7 @@ and for a single,
 </div>
 
 The Contributor is who submitted the mesh. For now Contributor. Description are in comments. If someone knows how to add captions to the plots, and acknowledgements for
-who contributed them, that would be nice to have those added and use these for the caption.
+who contributed them, that would be nice to have those added and use these for the caption. See below for the true formatting of the comments for now.
 
 At some point these figures should be moved out of the repo to where the other figures are located.
 
@@ -30,72 +30,72 @@ Please submit interesting contributions!
 
 <div align="center">
   <img src="https://github.com/trixi-framework/HOHQMesh_examples/blob/9df4021e865fcdb600a8a25ce2a37938a1e99523/gallery/cylinder-flow-small.png?raw=true" width="48%" />
-  [Contributor?. Cylinder with a wake region]: #
+	  <!---Contributor?. Cylinder with a wake region]: #-->
   <img src="https://github.com/trixi-framework/HOHQMesh_examples/blob/a8610caf1c5047741ee341462cb3e2f9d34471c6/gallery/naca0012-small.png?raw=true" width="48%" />
-  [David A. Kopriva. NACA0012 Airfoil]:
+  <!---[David A. Kopriva. NACA0012 Airfoil]:-->
 </div>
 
 <div align="center">
   <img src="https://github.com/trixi-framework/HOHQMesh_examples/blob/a8610caf1c5047741ee341462cb3e2f9d34471c6/gallery/three-element-30p30n-small.png?raw=true" width="48%" />
-   [Andrew R. Winters. Three element Airfoil from http://www.ae.metu.edu.tr/tuncer/ae546/multi/]: #
+   <!---[Andrew R. Winters. Three element Airfoil from http://www.ae.metu.edu.tr/tuncer/ae546/multi/]: #-->
 <img src="https://github.com/trixi-framework/HOHQMesh_examples/blob/a8610caf1c5047741ee341462cb3e2f9d34471c6/gallery/naca6412-small.png?raw=true" width="48%" />
-  [Andrew R. Winters. NACA6412 Airfoil from https://github.com/cfsengineering/GMSH-Airfoil-2D]: #
+  <!---[Andrew R. Winters. NACA6412 Airfoil from https://github.com/cfsengineering/GMSH-Airfoil-2D]: #-->
 </div>
 
 ## Coastlines
 
 <div align="center">
   <img src="https://github.com/trixi-framework/HOHQMesh_examples/blob/a8610caf1c5047741ee341462cb3e2f9d34471c6/gallery/indian-ocean-small.png?raw=true" width="48%" />
-  [David A. Kopriva. Indian Ocean]: #
+  <!---[David A. Kopriva. Indian Ocean]: #-->
   <img src="https://github.com/trixi-framework/HOHQMesh_examples/blob/a8610caf1c5047741ee341462cb3e2f9d34471c6/gallery/SuperiorFull-small.png?raw=true" width="48%" />
-  [David A. Kopriva. Lake Superior]: #
+  <!---[David A. Kopriva. Lake Superior]: #-->
 </div>
 <div align="center">
   <img src="https://github.com/trixi-framework/HOHQMesh_examples/blob/a8610caf1c5047741ee341462cb3e2f9d34471c6/gallery/east-coast-us-2d-small.png?raw=true" width="48%" />
-  [Joe Schoonover. East coast of the US]: #
+  <!---[Joe Schoonover. East coast of the US]: #-->
[…]
