### 2024-11-13 1d82a2e
Add captions to figures in gallery
