### 2024-11-13 5efb82f
Explain how to add new gallery examples
