### 2024-11-13 #7418 @fhindenlang
In principle yes... how much time do I have to review the paper?
