### 2024-11-13 669c482
adjust docs and examples readme to match in descriptions and order
