+++ Documentation/docs/examples.md
@@ -18,14 +18,14 @@ that can (like any of the examples) be modified at will.
 | Half Circle  | A half circle  |:material-check: | :material-close: | :material-close: | :material-check: | :material-close: | :material-check: | :material-close: | :material-close: | :material-close:| :material-close:|
 | Indian Ocean  | Complex domain with islands and inlets  |:material-check: | :material-check:3 | :material-close: | :material-close: | :material-check: | :material-close: | :material-close: | :material-close: | :material-close:| :material-close:|
 | KT3Element  | Three element Karman-Trefftz airfoil  |:material-close: | :material-check:3 | :material-close: | :material-close: | :material-check: | :material-close: | :material-close: | :material-close: | :material-close:| :material-close:|
-| Malpasset  | Scene of the Malpasset Dam failure  |:material-check: | :material-close: | :material-close: | :material-close: | :material-check: | :material-close: | :material-close: | :material-close: | :material-close:| :material-close:|
+| Lake Superior  | Complex domain with islands  |:material-check: | :material-check:2 | :material-close: | :material-close: | :material-check: | :material-close: | :material-close: | :material-close: | :material-close:| :material-close:|
+| Malpasset  | Region of Malpasset Dam failure  |:material-check: | :material-close: | :material-close: | :material-close: | :material-check: | :material-close: | :material-close: | :material-close: | :material-close:| :material-close:|
 | NACA0012  | Standard Airfoil geometry  |:material-close: | :material-check: | :material-close: | :material-check: | :material-close: | :material-close: | :material-close: | :material-close: | :material-check:| :material-close:|
 | PacMan  | A symmetric mesh  |:material-check: | :material-close: | :material-close: | :material-check: | :material-close: | :material-check: | :material-close: | :material-close: | :material-close:| :material-close:|
 | Pill  | Oblong domain with interior circles  |:material-check: | :material-check:3 | :material-check: | :material-close: | :material-close: | :material-close: |:material-close: | :material-close: | :material-close:| :material-close:|
 | RotatedEllipseWithSubregions  | Four circles, two half ellipses, and one quarter ellipse enclosed by an ellipse |:material-check: | :material-check:4 | :material-close: | :material-check: | :material-close: | :material-close: | :material-close: | :material-check: | :material-close:| :material-close:|
 | SplineGeometry  | Free form domain defined as a spline  |:material-check: | :material-close: | :material-close: | :material-close: | :material-check: | :material-close: |:material-close: | :material-close: | :material-close:| :material-close:|
 | Square  | Generation of a Cartesian mesh with no model  |:material-close: | :material-close: | :material-close: | :material-close: | :material-close: | :material-close: | :material-close: | :material-close: | :material-close:| :material-close:|
-| SuperiorMain  | Coastline of Lake Superior  |:material-check: | :material-close: | :material-close: | :material-close: | :material-check: | :material-close: | :material-close: | :material-close: | :material-close:| :material-close:|
 
 
 
