### 2024-11-13 39751b1
Update Gallery.md
Change all the div's to be the css versions for two column display plus captions. Change the captions to be of the form `caption (contributor)'. Change the video section name to "Applications".
