+++ Documentation/docs/Gallery.md
@@ -2,23 +2,35 @@
 Interesting meshes can be added below, with new categories added as needed. A two-column
 format is used, so to add two plots, use
 
-<div align="center">
-  <img src="https://github.com/trixi-framework/HOHQMesh_examples/blob/a8610caf1c5047741ee341462cb3e2f9d34471c6/gallery/meshPlotLeft.png" width="48%" />
-   [Contributor. Description]: #
- <img src="https://github.com/trixi-framework/HOHQMesh_examples/blob/a8610caf1c5047741ee341462cb3e2f9d34471c6/gallery/meshPlotright.png" width="48%" />
-   [Contributor. Description]: #
+<div class="gallery-two-columns">
+  <div class="gallery-column">
+    <figure>
+      <img src="File1ToInclude.png?raw=true"/>
+      <figcaption>
+         Caption (Contributor)
+       </figcaption>
+    </figure>
+  </div>
+
+  <div class="gallery-column">
+    <figure>
+      <img src="File2ToInclude.png?raw=true"/>
+      <figcaption>
+        Caption (Contributor)
+      </figcaption>
+  </div>
 </div>
 
 and for a single,
-<div align="left">
-  <img src="https://github.com/trixi-framework/HOHQMesh_examples/blob/a8610caf1c5047741ee341462cb3e2f9d34471c6/gallery/meshPlot.png" width="48%" />
-   [Contributor. Description]: #
+<div class="gallery-column">
+   <figure>
+      <img src="FileToInclude.png?raw=true"/>
+      <figcaption>
+        Caption (Contributor)
+      </figcaption>
 </div>
 
-The Contributor is who submitted the mesh. For now Contributor. Description are in comments. If someone knows how to add captions to the plots, and acknowledgements for
-who contributed them, that would be nice to have those added and use these for the caption. See below for the true formatting of the comments for now.
-
-At some point these figures should be moved out of the repo to where the other figures are located.
+The Contributor is who submitted the mesh. See below for where to submit figure contributions.
 
 -->
 
@@ -31,26 +43,12 @@ and then updating the gallery page
 
 ## Exterior Domains
 
-<div align="center">
-  <img src="https://github.com/trixi-framework/HOHQMesh_examples/blob/9df4021e865fcdb600a8a25ce2a37938a1e99523/gallery/cylinder-flow-small.png?raw=true" width="48%" />
-	  <!---Contributor?. Cylinder with a wake region]: #-->
-  <img src="https://github.com/trixi-framework/HOHQMesh_examples/blob/a8610caf1c5047741ee341462cb3e2f9d34471c6/gallery/naca0012-small.png?raw=true" width="48%" />
-  <!---[David A. Kopriva. NACA0012 Airfoil]:-->
-</div>
-
-<div align="center">
-  <img src="https://github.com/trixi-framework/HOHQMesh_examples/blob/a8610caf1c5047741ee341462cb3e2f9d34471c6/gallery/three-element-30p30n-small.png?raw=true" width="48%" />
-   <!---[Andrew R. Winters. Three element Airfoil from http://www.ae.metu.edu.tr/tuncer/ae546/multi/]: #-->
-<img src="https://github.com/trixi-framework/HOHQMesh_examples/blob/a8610caf1c5047741ee341462cb3e2f9d34471c6/gallery/naca6412-small.png?raw=true" width="48%" />
[…]
