### 2024-11-13 7f13575
Update Gallery.md
Fix a reference/caption
