### 2024-11-13 a5f36e0
Add Malpasset control file (#115)
* add Malpasset control file for reproducibility purposes

* Update README.md

Update documentation to include new elliptic arc and Malpasset examples.

* Update README.md

Alphabetize the list of names

* adjust docs and examples readme to match in descriptions and order

* make formatting of spline data nicer in Malpasset file

* replace Lake Superior control file and plot file with updated version

* update readme and docs

* update link for AllFeatures figure in the introduction portion of the docs

---------

Co-authored-by: David Kopriva <kopriva@math.fs
[…]
