+++ Examples/README.md
@@ -1,6 +1,6 @@
 # HOHQMesh Examples
 
-There are sixteen two-dimensional and ten three-dimensional examples included in the repository
+There are nineteen two-dimensional and ten three-dimensional examples included in the repository
 both for the user to run and to visualize models and meshes that use the features of HOHQMesh,
 and for developers to compare changes as algorithms are added or changed.
 
@@ -21,11 +21,13 @@ For this reason, we have reduced the storage requirements in the following ways:
    - HalfCircleArc
    - HalfCircleExtruded
    - HalfCircleRotated
+   - Malpasset
    - NACA0012
+   - PacMan
    - Pill
    - Pond3D
+   - RotatedEllipseWithSubregions
    - SplineGeometry
-   - PacMan
 
    Those that do not have a counterpart have plot file examples in this directory.
 
@@ -45,25 +47,27 @@ For this reason, we have reduced the storage requirements in the following ways:
 The table below provides a short summary of each two-dimensional example as well as a listing of the different features
 they exercise.
 
-| **Name** | **Description** | **Outer Boundary** | **Inner boundaries** | **Parametric Equation** | **Circular Arc** | **Spline** | **Line** | **Refinement Center** | **Refinement Line** | **Multiple Materials** |
-| :--- | :--- | :---: | :---: | :---: | :---: | :---: | :---: | :---: | :---: | :---: |
-| **All Features**          | Toy example; exercises all features                | ✅ | ✅ | ✅ | ✅ | ✅ | ✅ | ✅ | ✅ | ❌ |
-| **Bone and Marrow**       | Circular domain with two embedded boundaries       | ✅ | ✅ | ✅ | ❌ | ❌ | ❌ | ❌ | ❌ | ✅ |
-| **Box with refinement**   | Square domain with refinement regions              | ❌ | ❌ | ❌ | ❌ | ❌ | ❌ | ✅ | ✅ | ❌ |
-| **Cavity Ramp**           | Cavity domain with a sloping ramp on the exit side | ✅ | ❌ | ❌ | ❌ | ❌ | ✅ | ❌ | ❌ | ❌ |
-| **Circles3**              | Two circles enclosed by a large circle             | ✅ | ✅ | ✅ | ❌ | ❌ | ❌ | ❌ | ❌ | ❌ |
-| **East Coast US**         | Portion of US East coast from NYC to Jacksonville  | ✅ | ❌ | ❌ | ❌ | ✅ | ✅ | ❌ | ❌ | ❌ |
-| **Ellipse and 4 Circles** | Four circles enclosed by an ellipse                | ✅ | ✅ | ✅ | ❌ | ❌ | ❌ | ❌ | ❌ | ❌ |
-| **Gingerbread Man**       | Geometry with lots of holes                        | ✅ | ✅ | ✅ | ❌ | ✅ | ❌ | ❌ | ❌ | ❌ |
-| **Half Circle**           | Upper half circle                                  | ✅ | ❌ | ✅ | ❌ | ❌ | ✅ | ❌ | ❌ | ❌ |
-| **Indian Ocean**          | Complex domain with islands and inlets             | ✅ | ✅ | ❌ | ❌ | ✅ | ❌ | ❌ | ❌ | ❌ |
[…]
