### 2024-11-14 42bf190
delete gallery images (moved to examples repo)
