+++ Documentation/docs/hohqmesh-gallery/short-descriptions.md
@@ -1,25 +1,0 @@
-
-# "Useful" 2D meshes
-
-- cavity-ramp: Highlights that the spectral element mesh has big elements
-- cylinder-flow: Obstacle with a wake region
-- embedded-boundaries: Internal boundaries aligned
-- naca0012: Standard airfoil and mesh is nearly symmetric
-- naca6412: Another airfoil with an 11 degree angle of attack. Original data taken from https://github.com/cfsengineering/GMSH-Airfoil-2D
-- three-element-30p30n: Three element airfoil with many sharp features to capture. I got the data from this site http://www.ae.metu.edu.tr/tuncer/ae546/multi/ but there are many AIAA papers that use this geometry
-- lake-superior: Coastline data
-- east-coast-us: Coastline data and many scales
-- malpasset: Multiscale and coastline data. Far fewer elements than the FEM counterpart
-- indian-ocean: Multiple scales and boundaries captured
-
-# Fun 2D meshes
-- ghost
-- jack-o-lantern
-- julia-letters
-- gingerbreadman
-- snowman
-- trixi-hexe
-
-# Extruded 3D meshes with bottom data
-- east-coast-us-3d: Extrusion from given bathymetry data with a sea shelf
-- mt-st-helens: Automatic sizing with topography data, two figures neither of which I am 100% happy with
