### 2024-11-14 d4a68d9
AddGallery (#114)
* AddGallery

Add a Gallery page with plots.

* Update Gallery.md

Fix misspelling

* GalleryReferences

Add references to the gallery in the README and to the menu that shows up on the left. I also rearranged the menu to have the authors and license at the end.

* fix preparedocs

* use gallery images from examples repo

* suggest another formatting and video

* responsive video

* use smaller versions of the images to decrease latency

* Update Gallery.md

Fix multiple the column issue by including html comment markers in for the comments in the div's.

* Add captions to fi
[…]
