+++ Documentation/docs/Gallery.md
@@ -1,0 +1,255 @@
+<!---
+Interesting meshes can be added below, with new categories added as needed. A two-column
+format is used, so to add two plots, use
+
+<div class="gallery-two-columns">
+  <div class="gallery-column">
+    <figure>
+      <img src="File1ToInclude.png?raw=true"/>
+      <figcaption>
+         Caption (Contributor)
+       </figcaption>
+    </figure>
+  </div>
+
+  <div class="gallery-column">
+    <figure>
+      <img src="File2ToInclude.png?raw=true"/>
+      <figcaption>
+        Caption (Contributor)
+      </figcaption>
+  </div>
+</div>
+
+and for a single,
+<div class="gallery-column">
+   <figure>
+      <img src="FileToInclude.png?raw=true"/>
+      <figcaption>
+        Caption (Contributor)
+      </figcaption>
+</div>
+
+The Contributor is who submitted the mesh. See below for where to submit figure contributions.
+
+-->
+
+# A HOHQMesh Mesh Gallery
+
+Please submit interesting contributions by uploading images to
+[this repo](https://github.com/trixi-framework/HOHQMesh_examples)
+and then updating the gallery page
+[here](https://github.com/trixi-framework/HOHQMesh/blob/main/Documentation/docs/Gallery.md)!
+
+## Exterior Domains
+
+<div class="gallery-two-columns">
+  <div class="gallery-column">
+    <figure>
+      <img src="https://github.com/trixi-framework/HOHQMesh_examples/blob/9df4021e865fcdb600a8a25ce2a37938a1e99523/gallery/cylinder-flow-small.png?raw=true"/>
+      <figcaption>
+         Cylinder with a wake region (Andrew R. Winters)
+       </figcaption>
+    </figure>
+  </div>
+
+  <div class="gallery-column">
+    <figure>
+      <img src="https://github.com/trixi-framework/HOHQMesh_examples/blob/a8610caf1c5047741ee341462cb3e2f9d34471c6/gallery/naca0012-small.png?raw=true"/>
+      <figcaption>
+        NACA0012 Airfoil (David A. Kopriva)
+      </figcaption>
+  </div>
+</div>
+
+<div class="gallery-two-columns">
+  <div class="gallery-column">
+    <figure>
+      <img src="https://github.com/trixi-framework/HOHQMesh_examples/blob/a8610caf1c5047741ee341462cb3e2f9d34471c6/gallery/three-element-30p30n-small.png?raw=true"/>
+      <figcaption>
+         <a href="http://www.ae.metu.edu.tr/tuncer/ae546/multi/" target="_blank" rel="noopener noreferrer">Three element Airfoil</a> (Andrew R. Winters)
+       </figcaption>
+    </figure>
+  </div>
+
+  <div class="gallery-column">
+    <figure>
+      <img src="https://github.com/trixi-framework/HOHQMesh_examples/blob/a8610caf1c5047741ee341462cb3e2f9d34471c6/gallery/naca6412-small.png?raw=true"/>
+      <figcaption>
+        <a href="https://github.com/cfsengineering/GMSH-Airfoil-2D" target="_blank" rel="noopener noreferrer">NACA6412 Airfoil</a> (Andrew R. Winters)
+      </figcaption>
+  </div>
+</div>
+
+## Coastlines
+
+<div class="gallery-two-columns">
+  <div class="gallery-column">
+    <figure>
+      <img src="https://github.com/trixi-framework/HOHQMesh_examples/blob/a8610caf1c5047741ee341462cb3e2f9d34471c6/gallery/indian-ocean-small.png?raw=true"/>
[…]
