### 2024-11-14 05e6db4
Update README.md
Update link in the README to point to the docs page rather than the raw markdown file.
