### 2024-11-28 #7476 @fhindenlang
> @fhindenlang, @anderslogg can you please update us on the progress of your review?

Hi! I am also quite busy, but I started going through the checklist.
I did not yet have the time to install and test the package.
Some things that could be checked already:
* since the paper contains 'original data', there should be something like a zenodo link where an input listing/file for generating that mesh is available and also inputs/files that are needed for the simulation, if possible.
* I saw that some automated test runs were done, but I had difficulties finding out if meshes wqere actually generated in these runs. maybe it could be pointed out in the paper some of the tests that are running, and show them in the paper (and then add some input file in the zenodo link)...
