### 2024-11-30 d97af76
Update paper.md
Remove flow computation figure. Add links to the Malpasset control file, the gallery and the examples. Mention that meshes are generated as part of the test operation.
