+++ paper.md
@@ -78,16 +78,16 @@ HOHQMesh features include:
 
 HOHQMesh is available as an open-source software package under the MIT license and runs on Linux, macOS, and Windows [@kopriva2024hohqmesh].
 
-# Examples
+# Example
 
 In 1959 the Malpasset dam in France failed and flooded the Reyran river valley down to the Mediterranean sea [@hervouet1999malpasset],[@goutal1999malpasset]. Fig \ref{fig:Total} shows a mesh of the valley and a portion of the Mediterranean with 2392 fourth order elements generated by HOHQMesh in 0.44s on an Apple MacBook Pro with a 2.3 GHz Quad-Core Intel
-i7. A zoom of the western portion of the mesh is shown in Fig. \ref{fig:zoom}. The geometry model consists only of an outer boundary, which was specified as a cubic spline, and no inner boundaries. A control file to generate this mesh can be found in the [Examples/2D/Malpasset](https://github.com/trixi-framework/HOHQMesh/tree/main/Examples/2D/Malpasset) directory.
+i7. A zoom of the western portion of the mesh is shown in Fig. \ref{fig:zoom}. The geometry model consists only of an outer boundary, which was specified as a cubic spline, and no inner boundaries. Fig. \ref{fig:WaterHeight} shows a spectral element computation of the water heights using the mesh of Fig. \ref{fig:Total} in the package TrixiShallowWater.jl [@TrixiSW], which is part of the Trixi.jl [@schlottkelakemper2021purely] ecosystem.
 
 ![Spectral element mesh for the Reyran river valley including a portion of the Mediterranean Sea\label{fig:Total}](PaperFigs/malpasset-total-skeleton.png)
 
 ![Western portion of the Reyran valley mesh\label{fig:zoom}](PaperFigs/malpasset-total-skeleton-zoom.png)
 
-Further examples can be seen in the [gallery](https://trixi-framework.github.io/HOHQMesh/Gallery/). Control files for 29 examples that collectively include all the capabilities of the mesher are available in the [Examples](https://github.com/trixi-framework/HOHQMesh/tree/main/Examples) directory, a table of which is presented in the documentation under [Pre-Made Examples](https://trixi-framework.github.io/HOHQMesh/examples/). Finally, 21 meshes are generated and validated as part of the automated test option.
+![Spectral element computation of the water heights at 1985s after the break of the Malpasset dam\label{fig:WaterHeight}](PaperFigs/malpasset_time_1985s.png)
 
 # Related Software
 
