### 2024-11-30 356f19f
Update paper.md
Add flow computation back in.
