### 2024-12-02 0a239e1
Update the-model.md
