+++ .github/workflows/ci.yml
@@ -211,7 +211,7 @@ jobs:
         run: |
           sudo apt-get install -y lcov
           FC=${{ matrix.compiler }} ./Utilities/createcoverage
-      - uses: codecov/codecov-action@v4
+      - uses: codecov/codecov-action@v5
         if: ${{ matrix.coverage }}
         with:
           files: ./lcov.info
