### 2024-12-02 #7476 @fhindenlang
> We have added a link to the control file with which to generate the Malpasset mesh shown, and added a paragraph with links to a mesh gallery and the examples directory. There are 21 benchmark meshes that are generated as part of the test platform. That is now noted in the manuscript. The precise tests are listed as part of the test output, along with success or failure. (These tests are also done automatically upon a pull request.) A developer page exists for how to construct and include a test mesh, but that probably won't be of interest to users vs developers.

@DavidAKopriva  I might be mistaken, I just asked the editorialbot to regenerate the pdf (see above comment), and I do not see the changes you did in the pdf document.
@xuanxu I do not see a versioning of the document pdfs. the pdf that was generated on november 13 is now replaced by the pdf by the last comment, no way for me to tell if there are differences, I did not download the old version back then.
But since I cannot find the changes by David, I guess this is still the old version??

So how should we as reviewers check changes in the document?
