### 2024-12-02 #7476 @fhindenlang
OK, I found in the actions of the github repository the last version of the paper (https://github.com/trixi-framework/HOHQMesh/actions/runs/12098799170)
So this answers my last question to david.
@xuanxu I have not been given any information on how the process of checking changes in the document is supposed to work...
