### 2024-12-02 #7476 @fhindenlang
OK, so I was able to install two out of three ways:
[1] spack failed (kubuntu 22.04 linux system)
[2] downloading the tarball and installing with `make` and running the tests was sucessfull. I also ran the example from the paper.
[3] installation via julia (version 1.11.1, described here: https://github.com/trixi-framework/HOHQMesh.jl) was easy, and running the cases also. Unfortunately, the `GLMakie` package failed to install within julia, so I could not test the visualization part.

The process [2,3] generate a `.mesh` file. Is there any way how the mesh can be visualized?
