### 2024-12-02 #7476 @sloede
@fhindenlang Thanks a lot for your feedback!

> [1] spack failed (kubuntu 22.04 linux system)

Would you mind [creating an issue in the HOHQMesh repo](https://github.com/trixi-framework/HOHQMesh/issues/new) and letting us know what you tried and how it failed?

> [3] installation via julia [...] was easy, and running the cases also. Unfortunately, the GLMakie package failed to install within julia, so I could not test the visualization part.

Same here: Could you create an issue [here](https://github.com/trixi-framework/HOHQMesh.jl/issues/new) and give us the commands you used + the error logs? I just tried to run it on my Macbook and it seemed to work, but we'd be happy to look into it.
