### 2024-12-02 #7476 @xuanxu
> @xuanxu I do not see a versioning of the document pdfs. I have not been given any information on how the process of checking changes in the document is supposed to work...

@fhindenlang You are right. The idea is to review the latest version of the paper until the reviewers are happy with it. The papers are short (around 1000 words) but there is no versioning of them. If you have doubts and want to keep track of the changes, the only way I can think about is to check the code in the original repository (in this case [here's the list of recent commits](https://github.com/trixi-framework/HOHQMesh/commits/paper-2024-joss/) in the branch containing the paper, clicking in any commit will highlight the changes made there).
