### 2024-12-02 7ff543e
Response to reviewers' comments (#117)
* Update paper.md

Remove flow computation figure. Add links to the Malpasset control file, the gallery and the examples. Mention that meshes are generated as part of the test operation.

* Update paper.md

Add flow computation back in.

---------

Co-authored-by: David Kopriva <kopriva@math.fsu.edu>
