### 2024-12-02 #7476 @andrewwinters5000
@fhindenlang We updated the `paper-joss` branch of the repo and now the latest version of the pdf (compiled above) contains changes related to your comments.
