### 2024-12-03 #7476 @fhindenlang
> @fhindenlang Thanks a lot for your feedback!
>
> > [1] spack failed (kubuntu 22.04 linux system)
>
> ~Would you mind [creating an issue in the HOHQMesh repo](https://github.com/trixi-framework/HOHQMesh/issues/new) and letting us know what you tried and how it failed? Alternatively, try leaving out the `@main` part of the Spack install and see if that works - if yes, I'll fix the docs.~ **EDIT:** I've updated the spack repo with the latest versions of HOHQMesh, including a recipe for `main`. Thus the docs should work now 🤞 (otherwise please let me know).
>
> > [3] installation via julia [...] was easy, and running the cases also. Unfortunately, the GLMakie package failed to install within julia, so I could not test the visualization part.
>
> Same here: Could you create an issue [here](https://github.com/trixi-framework/HOHQMesh.jl/issues/new) and give us the commands you used + the error logs? I just tried to run it on my Macbook and it seemed to work, but we'd be happy to look into it.

Hi michael,
so somehow, the GLMakie can now be installed, I did not do anything different, and it worked! So I did not open an issue on the julia side.
For the spack I opened an issue in the repository, I just tried it and it still does not work. However, I was able to install it via `make` from  the downloaded release tarball, and could run the examples on my system.

So that completes my review, no more changes necessary from my side.
Very impressive work, a simple and fast mesh generator.
I would like to point out that the package is well documented and well maintained!
