### 2024-12-03 #7476 @DavidAKopriva
Thank you Florian for your comments and for all the (unique and thorough) effort it takes to review for JOSS. Getting an outside (re-)view and comments is really useful.
