### 2024-12-03 #7476 @andrewwinters5000
Many thanks @fhindenlang for your thorough review of the HOHQMesh repo. and paper! It has helped improve both.
