### 2024-12-04 213d21f
Update HOHQMeshMain.f90
Add help to be printed with the -help option.
