### 2024-12-04 #7476 @fhindenlang
I also raised the point on the mesh visualization.

At least I could use the glmakie package along with Hoqhmesh in julia to see the mesh of the interactive example.

 the mesh output file is not a standard, but AFAIK there is still no common standard for high order meshes, there are some attempts in cgns to have high order elements represented, and also there is the gmsh format.
