### 2024-12-04 #7476 @DavidAKopriva
I'm not sure what the issue is, but clearly there is one. The user guide https://trixi-framework.github.io/HOHQMesh/the-control-input/ shows how the control file specifies the files that are output, including a plot file and the format of the plot file (with SEM nodes or without). It also notes that the files can be viewed using a plotting program such as Visit or Paraview (or tecplot for that matter). The first two are free and cross platform. The paper is probably not the place to reproduce the user manual, so I'm not sure where one should explain how to plot the results except through the manual. The julia package is just a front end tool to this package (it was originally called HQMTool, but somewhere along the way it got renamed). There is another tool recently developed that lets one use SVG files from inkscape to generate control files. But those are just conveniences and not part of HOHQMesh. It is true that there is no common mesh file format for high order meshes, and efforts on my part have led to "interest" but no real movement. The mesh file formats that HOHQMesh generate are described in detail, again, in the documentation, https://trixi-framework.github.io/HOHQMesh/TheISMMeshFileFormats/, so there should be no confusion on that.
