### 2024-12-04 #7476 @fhindenlang
Well, I think the question was if the meshes can be visualized, and you confirmed it.
In my case, I did not read carefully the documentation, to know that the `.tec` file can be read with the tecplot reader of paraview, but I can now confirm that it works.
You are absolutely right that one should read the documentation carefully. But since there is already a "Generating a mesh" section on the Home of the documentation, it might be helpful to mention that is is possible to output a visualization file along the meshfile, for example `.tec` files that can be read with the paraview Tecplot reader.
