### 2024-12-04 #7476 @fhindenlang
And just a comment enabling the help or `-h` flag  on the executable.
In one of my Fortran codes, I use CLA, which you can include into your code to handle the command-line arguments. Its free and the license is very open. It can be downloaded, see details here:
https://ingria.ceoas.oregonstate.edu/fossil/CLAF90/doc/tip/README.md
