### 2024-12-04 #7476 @DavidAKopriva
Thanks for the CLAF90 suggestion. We've already got that parsing in HOHQmesh. And I have already written the help output (it's in a PR at the moment). It only took a few moments. I just never got around to that and had forgotten about it completely.
