### 2024-12-05 9376f5f
Add documentation for `-help` argument
