### 2024-12-05 #7476 @anderslogg
I confirm that I can open the Tecplot file with Paraview and the results look good.

Even if this can be figured out by reading the manual in detail, I suggest that a sentence is added to the README in the "Generating a mesh" section. Explain what output is created by the simple example and how to view the generated mesh (open Tecplot file with Paraview). The first thing I want to do when testing a new software is to run the simplest example and verify that I get the expected output (like the mesh for the Malpasset example in the paper).
