### 2024-12-05 7e06b86
Update README.md
Co-authored-by: Hendrik Ranocha <ranocha@users.noreply.github.com>
