### 2024-12-05 82bfe48
Add CMake-based builds to README
