### 2024-12-05 13b1cab
Modify command line behavior
Now accepts -h, -help, --help for help. Prints out help when there is an error in the command line arguments.
