+++ Source/HOHQMeshMain.f90
@@ -147,6 +147,7 @@
 !////////////////////////////////////////////////////////////////////////
 !
       SUBROUTINE ReadCommandLineArguments(version, test, generateTest, controlFileName, path)
+         USE, INTRINSIC :: iso_fortran_env, only : OUTPUT_UNIT
          USE CommandLineReader
          USE ProgramGlobals
          IMPLICIT NONE
@@ -155,10 +156,12 @@
 !        Arguments
 !        ---------
 !
-         CHARACTER(LEN=DEFAULT_CHARACTER_LENGTH) :: controlFileName, path
-         CHARACTER(LEN=*)                        :: version
-         LOGICAL                                 :: test, generateTest
+         CHARACTER(LEN=DEFAULT_CHARACTER_LENGTH)     :: controlFileName, path
+         CHARACTER(LEN=*)                            :: version
+         LOGICAL                                     :: test, generateTest
+         INTEGER                                     :: argCount
 
+         argCount = COMMAND_ARGUMENT_COUNT()
 
          IF ( CommandLineArgumentIsPresent("-version") )     THEN
             PRINT *, "HOHQMesh Version ", version
@@ -166,67 +169,94 @@
          END IF
 
          IF ( CommandLineArgumentIsPresent("-help") )     THEN
-            CALL PrintHelpMessage()
+            CALL PrintHelpMessage(OUTPUT_UNIT)
+            STOP
+         END IF
+
+         IF ( CommandLineArgumentIsPresent("-h") )     THEN
+            CALL PrintHelpMessage(OUTPUT_UNIT)
+            STOP
+         END IF
+
+         IF ( CommandLineArgumentIsPresent("--help") )     THEN
+            CALL PrintHelpMessage(OUTPUT_UNIT)
             STOP
          END IF
 
          test = .false.
          IF ( CommandLineArgumentIsPresent("-test") )     THEN
             test = .true.
+            argCount = argCount - 1
          END IF
 
          generateTest = .false.
          IF ( CommandLineArgumentIsPresent("-generateTest") )     THEN
             generateTest = .true.
+            argCount = argCount - 1
          END IF
 
          printMessage = .false.
          IF ( CommandLineArgumentIsPresent("-verbose") )     THEN
             printMessage = .true.
+            argCount = argCount - 1
          END IF
 
          controlFileName = "none"
          IF ( CommandLineArgumentIsPresent(argument = "-f") )     THEN
             controlFileName = StringValueForArgument(argument = "-f")
+            argCount = argCount - 2
          END IF
 
          path = ""
          IF ( CommandLineArgumentIsPresent(argument = "-path") )     THEN
             path = StringValueForArgument(argument = "-path")
+            argCount = argCount - 2
          END IF
 
          IF ( CommandLineArgumentIsPresent(argument = "-sLimit") )     THEN
             maxLevelLimit = IntegerValueForArgument(argument = "-sLimit")
+            argCount = argCount - 2
+         END IF
+!
+!        ---------------------------------------
+!        See if there are any uncaught arguments
+!        ---------------------------------------
+!
+         IF(argCount .ne. 0)   THEN
+            WRITE(OUTPUT_UNIT,*) "Command line argument error"
+            CALL PrintHelpMessage(OUTPUT_UNIT)
[…]
