### 2024-12-05 e9501b4
UpdateHelp
Remove -h option. Add comment that partial command line aregumens are handled.
