### 2024-12-06 406f7e6
Allow macOS builds with CMake
