### 2024-12-06 dc09dbd
Add prerequisites
