+++ paper.md
@@ -48,7 +48,7 @@ HOHQMesh [@kopriva2024hohqmesh] generates unstructured all-quadrilateral and hex
 
 Spectral element methods (SEMs) use multiple degrees of freedom within elements to achieve high order accuracy and can be applied to complex geometries. Details of SEMs can be found in the books by Deville, Fischer and Mund [@Deville:2002fk], Karniadakis and Sherwin [@Karniadakis:2005fj], Hesthaven and Warburton [@Hestahven:1008th], and Kopriva [@Kopriva:2009nx].
 
-Open source spectral element packages now exist to compute solutions of a wide range of equations such as the compressible and incompressible Navier-Stokes, ideal and visco-resistive magnetohydodynamics, Euler gas dynamics, and shallow water equations, and include Nektar++ [@CANTWELL2015205], SemTex [@BLACKBURN2019106804], Sem2dPack [@ampuero2012spectral], SPECFEM [@Martire:2021xt], Nek5000 [@fischer2008nek5000], HORSES3D [@FERRER2023108700], FLEXI [@KRAIS2021186], FLUXO [@FLUXO], Trixi.jl [@schlottkelakemper2021purely; @ranocha2022adaptive], and NUMA [@Giraldo:2013bh].
+Open source spectral element packages now exist to compute solutions of a wide range of equations such as the compressible and incompressible Navier-Stokes, ideal and visco-resistive magnetohydrodynamics, Euler gas dynamics, and shallow water equations, and include Nektar++ [@CANTWELL2015205], SemTex [@BLACKBURN2019106804], Sem2dPack [@ampuero2012spectral], SPECFEM [@Martire:2021xt], Nek5000 [@fischer2008nek5000], HORSES3D [@FERRER2023108700], FLEXI [@KRAIS2021186], FLUXO [@FLUXO], Trixi.jl [@schlottkelakemper2021purely; @ranocha2022adaptive], and NUMA [@Giraldo:2013bh].
 
 The features of SEMs are now well-established. Like low order finite element methods, they can be applied to general geometries, but have exponential convergence in the polynomial approximation order. Discontinuous Galerkin (DGSEM) versions applied to hyperbolic problems have exponentially convergent dissipation and dispersion errors [@ainsworth2004], making them well suited for wave propagation problems. Discontinuous Galerkin SEMs are also especially suitable when material discontinuities are present. Approximations exist for high order quadrilateral/hexahedral and triangle/tetrahedral elements.
 
@@ -57,7 +57,7 @@ What some are now calling “classical” spectral element methods use tensor pr
 Unfortunately, unstructured meshes for quad/hex elements are difficult to generate even for low order finite elements [@Bommes:2013lc].
 The advantages not withstanding, a major impediment to the application of SEMs has been the availability of appropriate general purpose mesh generation software that can generate elements of arbitrary order, especially in open-source form.
 In 2002 Sherwin and Peiro [@Sherwin:2002vx] wrote:  "The development of robust unstructured high-order methods is currently limited by the inability to consistently generate valid computational meshes for complex geometries without user intervention." This has remained true particularly for quadrilateral and hexahedral meshes.
-For these reasons, HOHQMesh was developed to generate all-quadrilateral and extruded hexahedral meshes suitable for use with spectral element methods. HOHQMesh is a direct quadrilateral mesher, which generates quadrilateral elements by the subdivision method of Schneiders [@schneiders2000algorithms] rather than indirectly from a triangular mesh or by curving a low order mesh. It also sizes and curves the elements based on the length scales in the model, rather than try to modify an existing low order mesh.
[…]
