### 2024-12-06 49f2ceb
Add CMake-based builds and prerequisites to README (#123)
* Add CMake-based builds to README

* Small addendum

* Update README.md

Co-authored-by: Hendrik Ranocha <ranocha@users.noreply.github.com>

* Allow macOS builds with CMake

* Add prerequisites

* Fix formatting

---------

Co-authored-by: Hendrik Ranocha <ranocha@users.noreply.github.com>
Co-authored-by: Andrew Winters <andrew.ross.winters@liu.se>
