### 2024-12-06 #7476 @anderslogg
> Thanks a lot for your feedback, @anderslogg!
>
> > Fortran (gfortran) is not listed as a dependency. I don't normally have it on my system (Mac) but it was easy enough to `brew install gfortran`.
>
> We have added a list of prerequisites to the [README](https://github.com/trixi-framework/HOHQMesh#prerequisites), including a Fortran compiler.
>
> > The executable does not respond to `--help`. I expect to be able to inspect which command-line flags are available.
>
> `--help` (or `-help` or `-h`) now prints a list of available command line flags.
>
> Output
> > The build clutters the root directory with `.o` and `.mod` files. It would be better to place the build artifacts in a sub directory.
>
> We have documented how to use the CMake-based build in the [README](https://github.com/trixi-framework/HOHQMesh#using-cmake), which enables out-of-tree builds.
>
> > How do I view the generated mesh? I don't know how to view ISM files. I tried changing to Abaqus in the control file but still can't open it in Paraview.
>
> We have documented how to view the Tecplot files in Paraview in the [README](https://github.com/trixi-framework/HOHQMesh#generating-a-mesh).
>
> > Remarks on paper:
> > ```
> > 1. Typo: "magnetohydodynamics"
> >
> > 2. Weird grammar: [...]
> > ```
>
> Fixed, thank you. I will generate the PDF with the next comment.
>
> With this, we hope to have addressed all reviewers' concerns.

Yes, looks great. This addresses all my concerns and I'm happy now.
