### 2024-12-06 #7476 @sloede
@xuanxu As far as we can tell, both reviewers have completed the checklist and given their consent to the manuscript. Is there anything else for us to do now?
