### 2024-12-06 #7476 @fhindenlang
The currently latest release 1.5.3 has not yet the help dialogue, but I guess a new release will be the outcome of this review, anyways.
