### 2024-12-06 #7476 @sloede
> The currently latest release 1.5.3 has not yet the help dialogue, but I guess a new release will be the outcome of this review, anyways.

Exactly. We wanted to wait until the paper has been officially accepted and then immediately create a new release that includes all proposed changes.
