### 2024-12-06 #7476 @anderslogg
> fhindenlang, @anderslogg Both your checklists are complete, can you confirm you now recommend the paper for publication?

Yes
