### 2024-12-09 #7476 @xuanxu
@andrewwinters5000 @sloede: looks like we're close to being done here. Please merge [this PR](https://github.com/trixi-framework/HOHQMesh/pull/124) and give your own paper a final proof reading.
