### 2024-12-09 5b9634f
Update paper.md
Co-authored-by: Michael Schlottke-Lakemper <michael@sloede.com>
