+++ paper.md
@@ -63,7 +63,7 @@ Examples of meshes generated by HOHQMesh have been published in @winters2014high
 
 # Features
 
-HOHQMesh is designed to require minimal input from the user through the use of a control file. The model defines the geometry in terms of outer and inner boundary curves.
+HOHQMesh is designed to require minimal input from the user through the use of a control file. The model defines the geometry in terms of an outer and one or more inner boundary curves.
 
 HOHQMesh features include:
 
