### 2024-12-09 5f0c394
Small edits for JOSS review (#124)
* Small edits

* Update paper.md

Co-authored-by: Michael Schlottke-Lakemper <michael@sloede.com>

---------

Co-authored-by: Andrew Winters <andrew.ross.winters@liu.se>
Co-authored-by: Michael Schlottke-Lakemper <michael@sloede.com>
