### 2024-12-09 #7476 @andrewwinters5000
@xuanxu I have done the items in the list above and the author list, affiliations, and ORCIDs are good in both the paper and Zenodo, The license for the archive and software match. Below are the version number and Zenodo DOIs

> * Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.

The new release with the latest changes has been made. It is v1.5.4 (https://github.com/trixi-framework/HOHQMesh/releases/tag/v1.5.4)
> * Archive the release on Zenodo/figshare/etc and post the DOI here.

Here are the appropriate links
DOI (version-specific): https://doi.org/10.5281/zenodo.14342391
DOI (version-agnostic): https://doi.org/10.5281/zenodo.13959058
