+++ paper.bib
@@ -29,7 +29,7 @@
 	author = {Kopriva, David A and Winters, Andrew R and Schlottke-Lakemper, Michael and Ranocha, Hendrik},
 	doi = {10.5281/zenodo.13959071},
 	howpublished = {\url{https://github.com/trixi-framework/HOHQMesh.jl}},
-	title = {{HOHQM}esh.jl: A Julia frontend to the Fortran-based HOHQMesh mesh generator for high order elements},
+	title = {{HOHQM}esh.jl: A {J}ulia frontend to the Fortran-based HOHQMesh mesh generator for high order elements},
 	year = {2024},
 	bdsk-url-1 = {https://doi.org/10.5281/zenodo.13959071}}
 
@@ -91,7 +91,7 @@
 	doi = {10.1007/978-3-319-12886-3_8},
 	pages = {133--152},
 	publisher = {Springer},
-	title = {Mesh Curving Techniques for High Order Discontinuous Galerkin Simulations},
+	title = {Mesh Curving Techniques for High Order Discontinuous {G}alerkin Simulations},
 	year = {2015},
 	bdsk-url-1 = {https://doi.org/10.1007/978-3-319-12886-3_8}}
 
