### 2024-12-10 #7476 @danielskatz
👋 @andrewwinters5000 - I have a few small suggested changes in https://github.com/trixi-framework/HOHQMesh/pull/125 - please merge this or let me know what you disagree with.

Also, as @xuanxu said, please update the title of the Zenodo archive to be the same as the paper's. This is a metadata-only change, and you don't need to make a new deposit or get a new DOI.
