### 2024-12-10 069167e
minor changes for JOSS paper (#125)
* Update paper.md

* Update paper.bib
