+++ paper.md
@@ -56,7 +56,7 @@ What some are now calling “classical” spectral element methods use tensor pr
 
 Unfortunately, unstructured meshes for quad/hex elements are difficult to generate even for low order finite elements [@Bommes:2013lc].
 The advantages not withstanding, a major impediment to the application of SEMs has been the availability of appropriate general purpose mesh generation software that can generate elements of arbitrary order, especially in open-source form.
-In 2002 Sherwin and Peiro [@Sherwin:2002vx] wrote:  "The development of robust unstructured high-order methods is currently limited by the inability to consistently generate valid computational meshes for complex geometries without user intervention." This has remained true particularly for quadrilateral and hexahedral meshes.
+In 2002, @Sherwin:2002vx wrote:  "The development of robust unstructured high-order methods is currently limited by the inability to consistently generate valid computational meshes for complex geometries without user intervention." This has remained true particularly for quadrilateral and hexahedral meshes.
 For these reasons, HOHQMesh was developed to generate all-quadrilateral and extruded hexahedral meshes suitable for use with spectral element methods. HOHQMesh is a direct quadrilateral mesher, which generates quadrilateral elements by the subdivision method of Schneiders [@schneiders2000algorithms] rather than indirectly from a triangular mesh or by curving a low order mesh. It also adjusts the size and curvature of the elements based on the length scales in the model, rather than attempting to modify an existing low-order mesh.
 
 Examples of meshes generated by HOHQMesh have been published in @winters2014high, @kopriva2016geometry, @acosta2020simulation, @manzanero2020entropy, @ersing2024entropy, @ranocha2024robustness, @PhysRevFluids.9.053901, plus @wintermeyer2018 and @eriksson2024laplacian.
