### 2024-12-10 #7476 @sloede
@danielskatz Thank you for the review! We have merged your PR and updated the Zenodo title. TIL that you can modify Zenodo metadata without a new release 🙏
