### 2024-12-10 #7476 @danielskatz
👋 @andrewwinters5000 (or another co-author) - please confirm that the pdf this produces looks good to you
