### 2024-12-10 #7476 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
