### 2024-12-11 #7476 @andrewwinters5000
> 👋 @andrewwinters5000 (or another co-author) - please confirm that the pdf this produces looks good to you

The final proof looks good and is a ready to be published.
