### 2024-12-11 #7476 @danielskatz
@andrewwinters5000 - The DOI is not yet resolving for me, so I will keep this open until it does. Normally it does fairly quickly, and I don't see anything on the Crossref status page to indicate that they are having any problems, so the issue is a mystery for me now, but one that will either resolve itself or we can work to fix in a little while
